Add custom hosts file support to Dns
This PR teaches System.Net.Dns about a new DOTNET_SYSTEM_NET_HOSTS environment variable.  When set, Dns will load a hosts file at a path specified in the environment variable and will use that hosts file as part of queries.  For any queries that result in hits against that hosts file, only the data from the hosts file will be used; otherwise, as happens with the system-wide hosts file, it falls through to the OS functionality.