SPMI: HTML escape lines inside diffs
The summarize script will convert markdown diff code blocks into
manually colored HTML tags because AzDO does not support proper
highlighting in its preview. However, when we do this we should also
take care to escape the contents as otherwise it can start being
interpreted as HTML tags. For example, in a recent case a field named
"<Option>k__BackingField" caused the preview to be truncated because
<option> was interpreted as a HTML tag.