JIT: Support complex gtCallAddr for marshalled pinvokes
After #85349 it is possible that we see GT_FTN_ADDR here and gtClone
does not support cloning it. While we could support cloning GT_FTN_ADDR
it is unfortunate to have cross phase dependencies on these operands
being of certain shapes, so this makes morph's argument manipulations
more robust in the scenario.

Fix #87888