JIT: Optimize aggregate info iteration in physical promotion
In several places we have to iterate over all promoted aggregates. This
was less efficient than it has to be because we stored aggregate
information in a flat array indexed by local number, and the vast
majority of locals are not physically promoted. This change introduces
an AggregateInfoMap that stores just the promoted aggregates, and
additionally stores an array mapping locals to the promoted aggregates.

This does mean that we need a two level lookup to map a local number
back to its aggregate information, but that still seems better for
throughput than the previous behavior.