Block tiering more broadly if the debugger disables optimizations
In #88199 the debugger is overriding some jit flags, but has left other flags set
that confuse the jit: both `TIER0` and `DEBUG_CODE` end uip set so one part of the jit
used logic appropriate for `TIER0` and another part did not.

The JIT ended up creating a patchpoint for OSR in a method with localloc and this
combination is not supported by OSR and lead to a crash.

Adjust the runtime so that if optimizations are disabled for a module by the debugger
then none of the methods in the module are tiering eligible. This stops the runtime
from setting `TIER0` in this case.

Also harden the jit so if the runtime asks for debuggable code, the jit will never
generate patchpoints.