Bring back old array enumerator code
This is a 0.3% size saving for Stage1. We not only allow array enumerators to be preinitialized again, but also avoid introducing many array `MethodTables` (looks like the new enumerators force array MethodTables for cases where we could have avoided them).

I shaped the old code to look like after the refactor in #82751, but reintroduced the old classes.

Fixes #82993.