Migrate from realpath to fullpath
realpath is guaranteed to resolve symlinks. fullpath may resolve symlinks.
Moves all existing code to call fullpath, which is the existing contract. On
Unix, fullpath calls realpath, meaning it resolves symlinks. On Windows it
doesn't. Code which requires symlinks to be resolved should be moved to call
realpath. realpath is temporarily renamed realpath2 in this commit to find
dangling references.