Switch the order of AOT SDK and framework assemblies
When you have built a CoreCLR CoreLib, it "poisons" the build
s.t. running libraries tests with NAOT becomes impossible.

This is because $(IlcFrameworkPath) draws from a shared framework
that contains the CoreCLR CoreLib, while it should use the CoreLib
from the AOT SDK.

This targeted change changes the order in which these two assemblies
get passed to ILC and allows one to use "/p:TestNativeAot=true"
without additional workarounds.