Add an option to not generate precise GC info
Revived #75817 with a small change.

Follow up to #75803.

If enabled, conservative GC stack scanning will be used and metadata related to GC stack reporting will not be generated. The generated executable file will be smaller, but the GC will be less efficient (garbage collection might take longer and keep objects alive for longer periods of time than usual).

Saves 4.4% in size on a Hello World. Saves 6.7% on Stage 1. I'll take that.

The main difference from the previous PR is that I'm no longer dropping GC info on `UnmanagedCallersOnly` methods and that seems to be enough to make Smoke tests pass.