[mono] Align HOST_TVOS/TARGET_TVOS defines with the rest of the repo
In the mono build we've been setting HOST_IOS/TARGET_IOS when targetting tvOS which differs from the other runtime repo build systems.
This aligns the defines so we only set the iOS ones on iOS and not on tvOS.

Also removes mono_dl_get_system_dir() since it was not actually used.