Use SearchValues for non-ASCII Regex fallback
When we encounter a set for which we can't use one of the IndexOfXx variants, we use IndexOfAny with SearchValues.  If the whole set is ASCII, this is easy.  But if there are any non-ASCII characters in the set, we currently emit a helper method that first does a search for all of the ASCII values or anything that's non-ASCII, and then if something non-ASCII is found, it proceeds to do a scalar scan matching every character.

Now that SearchValues has a vectorized implementation of a probabilistic map, we can instead fall back to that when the set isn't too large and thus can be fully enumerated into a SearchValues instance.  We still do this as a two-step process, as searching for the ASCII subset is measurably faster than the probabilistic map search.  However, if we see that there's no ASCII at all in the target set, we can then skip the helper entirely and just do the IndexOfAny with the SearchValues for the non-ASCII targets.