Fix GC hole with `filter` liveness
Execution following a `filter` clause can be to finally/fault clauses nested within
the `try` region that the `filter` protects. E.g.,

```
   try {
      try {
      } finally {
         // 1
      }
   } filter {
      // 2
   } filter-handler {
   }
```

In this case, after `2` executes, `1` might execute. This is a consequence
of the 2-pass EH model. This means that GC variables in `1` must be kept alive
when `2` is executed, in case a GC occurs during the execution of `2`.

Code to handle this case was introduced with https://github.com/dotnet/coreclr/pull/23044,
however it didn't handle the case where a `filter` is a top-level EH construct
(and not nested within another `try`).

This change callers of `fgGetHandlerLiveVars` to always call it (not just
when `ehBlockHasExnFlowDsc()` is true) so the code that handles the filter case
has a chance to run. To make this a little more efficient, the VARSET_TP that will
be filled in is expected to be pre-allocated and passed in.

Fixes #88168