Trim non-essential comments.
This includes comments that:

1. Can become outdated if methods are renamed (see https://github.com/dotnet/runtime/pull/89100#discussion_r1267742108).

2. Are not absolutely critical (see https://github.com/dotnet/runtime/pull/89100#discussion_r1268482921).

I also removed the mention of using FileOptions.WriteThrough as an
alternative to RandomAccess.FlushToDisk(). This is an advanced scenario
that can actually degrade performance if done incorrectly (i.e. if the
user doesn't manage their own buffers such that all writes keep hitting
the disk) so it's probably not wise to mention it unless we go into more
detail than is ideal for inclusion in the Remarks section of an XML doc
comment.

To remove the lengthy comment in the ValidateInput() function I had to
rename the optional parameter from "flushingToDisk" to the more generic
and self-describing "allowUnseekableHandles" (see https://github.com/dotnet/runtime/pull/89100#discussion_r1267724547).