Disable networking in XML resolver in AOT compilers
Use the switch introduced in #84169. The AOT compilers don't need this. Speeds up the build by a couple seconds because we no longer compile garbage. Shrinks size of ilc.exe from 15.2 MB to 11.1 MB.

One the Native AOT side, this will only kick in once we update the SDK used to build the repo to something that has this (RC1?). On the crossgen2 side this takes effect immediately.