SPMI: Fix a couple of replay/record bugs (#89443)
* Avoid mutating constrained token in CEEInfo::getCallInfo

For certain calls the EE would modify the constrained token passed to
getCallInfo. This token is an in-argument, and SPMI does not expect it
to be modified. Fix the EE to make a local copy when it needs to mutate
it.

* SPMI: Fix treatment of CORINFO_FLG_BAD_INLINEE during replay

- Save union of method attributes set by the JIT from
  recSetMethodAttribs calls
- Fix bit handling in repGetMethodAttribs
- Add the same method attribs handling to repGetCallInfo that
  repGetMethodAttribs has

This fixes an issue where we could have recorded an old result of
getCallInfo without CORINFO_FLG_DONT_INLINE set. If the JIT then calls
repSetMethodAttribs(CORINFO_FLG_BAD_INLINEE) this would not be reflected
in subsequent calls to getCallInfo, causing possible misses.