Update description of DOTNET_JitStressProcedureSplitting (#89576)
When the JitStressProcedureSplitting mode was first implemented, we decided to
skip functions with exception handling, since we hadn't decided how to handle
cold EH funclets just yet. But for about a year now, JitStressProcedureSplitting has worked
with functions with EH funclets (see Compiler::fgDetermineFirstColdBlock), so its description
in jitconfigvalues.h should reflect this.