Add more verification in EventPipe rundownvalidation test.
Subtle changes like:
https://github.com/dotnet/runtime/commit/2f2aaae4e4fc1c33d6a7db316fcc0cac4df7af3f

Knocked out a couple of rundown events on Mono that was detected and
fixed in https://github.com/dotnet/runtime/pull/88634.

It would been ideal to catch this issue as part of CI, but turns out
that our existing rundownvalidation test only tests a couple of rundown
events, and not any of the module/assembly events that was affected by
above change.

This commit adds validation that the following additional rundown events
are included in the rundown event stream:

RuntimeStart
MethodDCStopInit
MethodDCStopComplete
LoaderModuleDCStop
LoaderDomainModuleDCStop
LoadAssemblyDCStop
LoadAppDomainDCStop