JIT: don't dead store untracked OSR locals based on ref count
The local var ref count for OSR locals can be misleading, since some
of the appearances may be in the "tier0" parts of the methods and
won't be visible once we trim the OSR method down to just the part
we're going to compile.

So, we can't rely on ref counts to enable a dead store of an OSR local.

Fixes #89666.