Don't generate symbols if user doesn't want them
This has two changes:

* Stop looking at `DebugSymbols` property. It doesn't look like the SDK actually looks at this (anymore?) either. Setting this to false still produces managed symbols. The only thing that works is setting `DebugType`.
* Don't generate DBG file if debugging symbols are not generated. The user expectation when debugging symbols are disabled is that they don't get the extra file on disk. The DBG file would still have debug information for the native runtime. We may or may not run objcopy to strip symbols.

I considered also running objcopy when StripSymbols is set to false but DebugType is set to none like we discussed in #88780 but maybe that goes too far - I'd expect users would only set StripSymbols to a non-default value in case they e.g. don't want to install objcopy and we'd need yet another property to say "no really don't run objcopy". So I made this orthogonal to that.