Pass through ComWrappers based objects to COM type descriptor
WinForms hosts the the COM type descriptor. It has been updated to work with ComWrappers based objects- adding the check for them in TypeDescriptor.

There are hard dependencies on Windows in the current implementation. We can potentially update to conditionalize these. https://github.com/dotnet/winforms/issues/9291 tracks.

There is no automated test as it would require adding a dependency upstream to WinForms.