[iOS] Generate unique class name for MonoDeadLetter
In library mode, it is possible to have multiple shared libraries (runtimes) loaded in the same process. We have a small bit of objc to make sure we know when threads die to we can properly detach the runtime. Unfortunately, the class name we use is not unique and results in a crash when calling into a 2nd library.

This change makes the class name partially unqiue to avoid such a circumstance.