Fix isIPAdjusted setting during exception handling (#89930)
* Fix isIPAdjusted setting during exception handling

The CrawlFrame::isIPAdjusted flag is set to true for all frames
when handling software exceptions. That is not correct, as the IP is not
adjusted in those cases.

I've also cleaned up the values the flag is set to at various places to
use the real type of the member, which is `bool`. It was being set to
TRUE / FALSE or even 0 / 1 at some places.

As a related cleanup, I've removed the FRAME_ATTR_OUT_OF_LINE flag as it
was never set anywhere and still was being tested at few places.

* Add regression test