[workloads] Use native arm64 aot compilers for iOS, tvOS, maccatalyst, and android where possible
As a result of https://github.com/dotnet/runtime/pull/89027 and https://github.com/dotnet/runtime/pull/74715, we can now build native arm64 cross compilers on linux and macos. This change adds the right references to the workload manifest so that they can be installed with maui and the mobile SDK's.

win-arm64 is the only arm64 platform left and that is a work in progress.

Contributes to https://github.com/dotnet/runtime/issues/82495

Fixes https://github.com/dotnet/maui/issues/4476