Improve test failure messages
I've noticed that there are more coding error scenarios where a test will fail and the failure that is hit is the failed to resolve "test.exe" error.  I don't fully have my head around which exceptions are being caught and logged but it seems like a lot more than there used to be.

Suffice it to say, when the only output from a test failure is about failing to resolve test.exe, that isn't particularly helpful when what actually happened is you had a mistake in your code and an exception was thrown.

With the linker now eating more exceptions and returning a non-zero exit code rather than throwing and crashing I think it's worth having the test framework fail with the messages from the linker if the exit code is non-zero.

For the tests that expect a non zero exit code, I added `ExpectNonZeroExitCodeAttribute`.  Currently there is only 1 test that needs this.