Remove spurious assertion check causing .NET Core crashes
I started this investigation based on heads-up from Rich Lander
who hit the issue in a PR run. Thanks to help from Jan Vorlicek
I have been able to repro the issue and I discussed it with
Tanner Gooding who proposed this fix - removing the superfluous
assertion check duplicating an upfront check in the method
NarrowUtf16ToAscii before calling into the intrinsified version.

I sincerely admit I'm not an expert in the field of instruction
set extensions and I have driven this investigation mostly because
David Wrighton is OOF on parental leave for the entire month of
August so I'm trying to make sure we ship .NET 8 in the best
possible quality; I'll be more than happy for any suggestions
for improvements or any other feedback.

Thanks

Tomas