Harmonize flushing of read-only files.
There is a difference in behaviour between FlushFileBuffers() on
Windows and fsync() on Unix when it comes to read-only files. On
Windows it fails with ERROR_ACCESS_DENIED while on Unix it just works.
We made a decision to harmonize the platforms by making Windows behave
like Unix (i.e. we catch and ignore the error if a user attempts to
flush a file opened for reading, so it is effectively a no-op).

See https://github.com/dotnet/runtime/pull/89100#discussion_r1283070967