SPMI: Capture EE API exceptions in getNewHelper, and simplify the API shape a bit (#89852)
We have various tests where getNewHelper throws an EE exception, so
capture these and rethrow them as part of SPMI.

It would be nice with a more general approach to this (conceptually all
EE APIs could throw exceptions), but this at least handles it for our
own libraries.

Contributes to #47540 and #47546.

Also clean up the API shape of getNewHelper slightly, since this requires
a JIT-EE GUID change anyway; the only used member of the token is the class,
so pass a class handle instead. The second parameter was also removed.